requirements.txt
numpy==2.1.2
Pillow==11.0.0
PyJWT==2.9.0
websockets==13.0.1
matplotlib==3.9.2